requirements.txt
# Core FastAPI Framework
fastapi==0.109.0
uvicorn==0.27.0

# Model Serving and Optimization
torch==2.2.0
transformers==4.38.0
deepspeed==0.12.0
vllm==0.3.1
tritonclient[all]==2.31.0

# Image Processing (for vision API)
opencv-python-headless==4.9.0.80
pillow==10.2.0

# Speech-to-Text
whisper==1.1.10
torchaudio==2.2.0

# Retrieval-Augmented Generation (RAG)
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Logging & Monitoring
loguru==0.7.2

# Security & Authentication
python-dotenv==1.0.1
pyjwt==2.8.0

# Database & Caching
redis==5.0.1
pymongo==4.6.2

# Testing & Benchmarking
pytest==8.0.0
pytest-asyncio==0.23.5
locust==2.18.0

# Miscellaneous
tqdm==4.66.1
requests==2.31.0
aiohttp==3.9.1
